Steps towards several fixes
 - FLO2D model call fail. Solution in preparation.

 - New confluences assignment dialog. Solution in preparation.

+++ flo2d/flo2d.py
@@ -564,17 +564,28 @@ class Flo2D(object):
         if not ok:
             return
         flo2d_dir, project_dir = dlg.get_parameters()
-
         if sys.platform != "win32":
             self.uc.bar_warn("Could not run simulation under current operation system!")
             return
-        try:
-            debugDAT = os.path.join(project_dir, "QGISDEBUG.DAT")
-            if os.path.exists(debugDAT):
-                os.remove(debugDAT)
-            simulation = FLOPROExecutor(flo2d_dir, project_dir)
-            simulation.run()
-            self.uc.bar_info("Simulation started!", dur=3)
+        try: 
+            return_code = -9999
+            if os.path.isfile(flo2d_dir + r"\FLOPRO.exe"):
+                debugDAT = os.path.join(project_dir, "QGISDEBUG.DAT")
+                if os.path.exists(debugDAT):
+                    os.remove(debugDAT)
+                simulation = FLOPROExecutor(self.iface, flo2d_dir, project_dir)
+                return_code = simulation.run()
+                if return_code != 0:
+                    self.uc.show_warn(
+                        "ERROR 190821.1120: FLO2D.EXE Model run failed!\n\n"
+                        + "Program finished with return code " + str(return_code)
+                    )
+                else:
+                    self.uc.show_info( "Model finished with return code "  + str(return_code))       
+                
+                            
+            else:
+                self.uc.show_warn("WARNING 180821.0841: Program FLO2D.exe is not in directory\n\n" + flo2d_dir)                
         except Exception as e:
             self.uc.log_info(repr(e))
             self.uc.bar_warn("Running simulation failed!")
